requirements.txt
keyboard
pystray
requests
Pillow
pyautogui
pygame
pywin32